% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoder.R
\name{enc_reqMktData}
\alias{enc_reqMktData}
\title{enc_reqMktData()}
\usage{
enc_reqMktData(
  self,
  contract,
  genericTicks = "100,101,104,106,165,221,225,236",
  snapshot = FALSE,
  regulatorySnaphsot = FALSE,
  mktDataOptions = list(),
  reqId = self$nextId()
)
}
\description{
enc_reqMktData()
}
