% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{handlers}
\alias{handlers}
\alias{hl_recorder}
\alias{hl_record_stdout}
\alias{hl_decode_str}
\title{IB Handlers}
\usage{
hl_recorder(file = stdout(), type = c("str", "val"), ts_format = TS_FORMAT)

hl_record_stdout(self, msg, ...)

hl_decode_str(self, msg, ...)
}
\arguments{
\item{file}{file to store the output, by default \code{\link[=stdout]{stdout()}}}

\item{ts_format}{format for timestamps}

\item{self}{\code{TWS} R client object created with \code{\link[=tws]{tws()}} constructor.}

\item{msg}{a list containing values added by the preceding handlers. The
initial message contains the following element
\describe{
\item{bin}{raw object containing the binary representation of the TWS message}
\item{ts}{timestamp when the message was received by TWS R client}
}}
}
\description{
When the \code{TWS} R client receives a message from the \code{TWS} server, each
messaged is processed in turn by the handlers in the handler pipeline
(\code{handlers} argument of \code{\link[=tws]{tws()}} function). Handlers are functions that take
two arguments - \code{TWS} R client object and a message. A handler should modify
and return the message. If a handler return NULL, the message is not passed
to the remaining handlers.
}
